requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
Flask-Cors==4.0.0
PyJWT==2.8.0
python-socketio==5.8.0
eventlet==0.33.3
bcrypt==4.0.1
PyYAML==6.0.1
python-dotenv==1.0.0
pathlib==1.0.1
